fix some visuals on Pet and Settings page

+++ src/pages/User/Profile/GraphsCarousel.js
@@ -54,7 +54,7 @@ export function GraphsCarousel(props) {
 
     for (let i = 5; i >= 0; i -= 1) {
       carouselData[0].data.unshift({
-        month: new Date(null, currentMonth - 1).toLocaleDateString("en", { month: "long" }),
+        month: new Date(null, currentMonth - 1).toLocaleDateString("en", { month: "short" }),
         data:
           totalData &&
           totalData[currentYear] &&
@@ -65,7 +65,7 @@ export function GraphsCarousel(props) {
       });
 
       carouselData[1].data.unshift({
-        month: new Date(null, currentMonth - 1).toLocaleDateString("en", { month: "long" }),
+        month: new Date(null, currentMonth - 1).toLocaleDateString("en", { month: "short" }),
         data:
           totalData &&
           totalData[currentYear] &&
@@ -76,7 +76,7 @@ export function GraphsCarousel(props) {
       });
 
       carouselData[2].data.unshift({
-        month: new Date(null, currentMonth - 1).toLocaleDateString("en", { month: "long" }),
+        month: new Date(null, currentMonth - 1).toLocaleDateString("en", { month: "short" }),
         data:
           totalData &&
           totalData[currentYear] &&
